nano: PerfectHash::build passes das::max two arguments of the same type
uint32_t is unsigned int on the hosted targets and unsigned long on arm-none-eabi, so
das::max(1u, (n + 4) / 5) deduces one template argument on a native build and two
different ones when nano cross-compiles for cortex-m4, where it stops compiling. Both
call sites now spell the literal uint32_t(1), the way every other das::max call in the
tree casts its literal side.

+++ src/simulate/name_lookup.cpp
@@ -55,8 +55,8 @@ namespace das {
 
     bool NameLookup::PerfectHash::build ( const vector<uint64_t> & keys, vector<uint32_t> & dispOut, uint32_t slack ) {
         uint32_t n = uint32_t(keys.size());
-        nbuckets = das::max(1u, (n + 4) / 5);
-        nslots = n + das::max(1u, n / 20) * slack;
+        nbuckets = das::max(uint32_t(1), (n + 4) / 5);
+        nslots = n + das::max(uint32_t(1), n / 20) * slack;
         vector<uint64_t> scrambled(n);
         vector<uint32_t> bucketOf(n);
         vector<uint32_t> bucketSize(nbuckets, 0);
